DMS-759 Make DMS Database Deploy Scripts Idempotent  (#657)
* DMS-759 Make DMS Database Deploy Scripts Idempotent

Update scripts to include IF EXISTS
Update insert to be idempotent

* Update scripts

Remove Create Publication process from Template
Remove env:NEED_DATABASE_SETUP validation

* Update scripts

Add validation to avoid running the pgdump restore again
Update claims Hierarchy insert to update or insert the value

* Fix scripts workflow

Update conditions and postgres scripts

* Update setup-database-template.psm1

Update message when data was loaded previously

* Update start-local-dms.ps1

Fix order for postgres container

* Update scripts to fix errors

* Update 0011_Insert_ClaimsHierarchy.sql

Fix claimset definition

* Fix script format

Replace string format by string concatenation
Fix Configure Replication

* Update 0010_Insert_Claimset.sql

Fix script to use the same order when inserting claimsets

* Update scripts to add alias to Subqueries

Add alias to subqueries

+++ eng/docker-compose/setup-database-template.psm1
@@ -39,14 +39,20 @@ function LoadSeedData {
 
     if ($LASTEXITCODE -eq 0) {
         Write-Host "PostgreSQL is running. Proceeding with bootstrap..."
+        $schemaName = "dms"
+        $schemaExists = docker exec dms-postgresql psql -U $User -d $Database -tAc "SELECT 1 FROM pg_namespace WHERE nspname = '$schemaName';"
+        if (-not $schemaExists) {
+            docker cp $sqlPath "dms-postgresql:/tmp/restore.sql"
+            docker exec dms-postgresql psql -U postgres -d $database -f /tmp/restore.sql
 
-        docker cp $sqlPath "dms-postgresql:/tmp/restore.sql"
-        docker exec dms-postgresql psql -U postgres -d $database -f /tmp/restore.sql
-
-        if ($LASTEXITCODE -eq 0) {
-            Write-Host "Template data loaded successfully."
-        } else {
-            Write-Error "Failed to execute the database script file."
+            if ($LASTEXITCODE -eq 0) {
+                Write-Host "Template data loaded successfully."
+            } else {
+                Write-Error "Failed to execute the database script file."
+            }
+        }
+        else {
+           Write-Warning "PostgreSQL Data Seed Load failed: existing volumes were detected with data loaded. Please remove them before proceeding."
         }
     } else {
         Write-Error "PostgreSQL server is not running or unreachable."
